Ingestion/requirements.txt
kafka-python
python-dotenv
logger
wasabi
forex-python
requests
rich
